NEW BRANCH + "CODE_OF_CONDUCT.md" Update
Creation of a new branch called "LTD ENV" which stands for - "Long-Term-Development Environment". This branch is isolated from the rest of T1's Continuous Deployment branches. Therefore this branch is isolated and should operate independently of the T1 web environments. Any updates to T1 code that are incomplete or currently non-functional should be committed to this LTD branch for further development without having any updates to the code affect T1's Deployment Quota

---
Additionally
---

The T1 Code of Conduct will be getting re-written. However, this will take some time and will be edited many times in the LTD ENV before any updates will be released to the Dev Env. or the Prod. Env.